+++ contracts/main/src/utils/mod.rs
@@ -1,10 +1,14 @@
 mod common;
 mod hash;
+mod merkle_tree;
+mod mmr;
 mod signature;
 
 pub use common::{
     check_aggregator, check_output_type_hash, fetch_capacities, load_action, load_global_state,
     CapacityChange,
 };
-pub use hash::{compute_account_root, new_blake2b, HashMerge};
+pub use hash::new_blake2b;
+pub use merkle_tree::{merkle_root, CBMT};
+pub use mmr::{compute_account_root, HashMerge};
 pub use signature::verify_ag_signature;
